hotfix transformsymlinks.sh

+++ transformsymlinks.sh
@@ -2,6 +2,7 @@ for FOLDER in data/*
 do
 for SCENE_FOLDER in $FOLDER/*
 do
+
 if [ ! -d "${SCENE_FOLDER}" ]; then 
 continue;
 fi
@@ -9,7 +10,7 @@ fi
 TRANSFORMS_PATH=$SCENE_FOLDER/transforms.json
 TRANSFORMS_TRAIN_PATH=$SCENE_FOLDER/transforms_train.json
 
-if ! [ -e $TRANSFORMS_PATH ]
+if [ ! -e $TRANSFORMS_PATH ]
 then
 echo "no transforms file $SCENE_FOLDER, trying to link"
 
@@ -17,9 +18,12 @@ if [ -L $TRANSFORMS_PATH ]; then
 unlink $TRANSFORMS_PATH
 fi
 
-if [ -f $TRANSFORMS_TRAIN ]
+if [ -e $TRANSFORMS_TRAIN_PATH ]
 then
+echo "linking transforms.json to ${TRANSFORMS_TRAIN_PATH}"
 ln -s transforms_train.json $TRANSFORMS_PATH
+else
+echo "${TRANSFORMS_TRAIN_PATH} does not exist"
 fi
 
 fi
